+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,6 +1,14 @@
+## 0.3.17+1
+
+* Converts main plugin tests to swift.
+
+## 0.3.17
+
+* Removes OCMock from tests.
+
 ## 0.3.16
 
-* Converts main plugin class to Swift
+* Converts main plugin class to Swift.
 
 ## 0.3.15